+++ ballcat-log/ballcat-log-model/src/main/java/com/hccake/ballcat/log/model/entity/AccessLog.java
@@ -12,8 +12,7 @@ import java.time.LocalDateTime;
 /**
  * 后台访问日志
  *
- * @author hccake
- * @date 2019-10-16 16:09:25
+ * @author hccake 2019-10-16 16:09:25
  */
 @Data
 @TableName("log_access_log")
@@ -40,7 +39,7 @@ public class AccessLog {
 	 * 用户ID
 	 */
 	@Schema(title = "用户ID")
-	private Integer userId;
+	private Long userId;
 
 	/**
 	 * 用户名
