+++ ballcat-system/ballcat-system-biz/src/main/java/com/hccake/ballcat/system/service/impl/SysRoleMenuServiceImpl.java
@@ -32,7 +32,7 @@ public class SysRoleMenuServiceImpl extends ExtendServiceImpl<SysRoleMenuMapper,
 	 */
 	@Override
 	@Transactional(rollbackFor = Exception.class)
-	public Boolean saveRoleMenus(String roleCode, Integer[] menuIds) {
+	public Boolean saveRoleMenus(String roleCode, Long[] menuIds) {
 		// 1、先删除旧数据
 		baseMapper.deleteByRoleCode(roleCode);
 		if (menuIds == null || menuIds.length == 0) {
@@ -72,7 +72,7 @@ public class SysRoleMenuServiceImpl extends ExtendServiceImpl<SysRoleMenuMapper,
 	 * @return 被更新的菜单数
 	 */
 	@Override
-	public int updateMenuId(Integer originalId, Integer menuId) {
+	public int updateMenuId(Long originalId, Long menuId) {
 		return baseMapper.updateMenuId(originalId, menuId);
 	}
