+++ ballcat-system/ballcat-system-model/src/main/java/com/hccake/ballcat/system/model/vo/SysMenuPageVO.java
@@ -24,13 +24,13 @@ public class SysMenuPageVO {
 	 * 菜单ID
 	 */
 	@Schema(title = "菜单ID")
-	private Integer id;
+	private Long id;
 
 	/**
 	 * 父级ID
 	 */
 	@Schema(title = "父级ID")
-	private Integer parentId;
+	private Long parentId;
 
 	/**
 	 * 菜单名称
